# visualizer/src/app/api/storage-api/model/newObject.ts
/**
 * Storage Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NewObject { 
    id?: string;
    name: string;
    imageUrl?: string;
}